# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Real state",
        "description": "API para gestionar propiedades inmobiliarias",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/solicitudes": {
            "get": {
                "tags": [
                    "SolicitudVisita"
                ],
                "summary": "Obtener todas las solicitudes de visita",
                "operationId": "086431570d6b2493541026e645665995",
                "responses": {
                    "200": {
                        "description": "Lista de solicitudes de visita",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/SolicitudVisita"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiKeyAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "SolicitudVisita"
                ],
                "summary": "Crear una nueva solicitud de visita",
                "operationId": "882472fbd0bd21050a21b1b0134128f5",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SolicitudVisita"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Solicitud de visita creada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SolicitudVisita"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Errores de validación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "errors": {
                                            "type": "object",
                                            "additionalProperties": {
                                                "type": "string"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error al guardar solicitud de visita",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Error al guardar solicitud de visita: mensaje del error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiKeyAuth": []
                    }
                ]
            }
        },
        "/api/solicitudes/{id}": {
            "get": {
                "tags": [
                    "SolicitudVisita"
                ],
                "summary": "Obtener una solicitud de visita por ID",
                "operationId": "038d064d6aef8c471de5df25f09bbe23",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la solicitud de visita",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Solicitud de visita encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SolicitudVisita"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Solicitud de visita no encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Solicitud de visita no encontrada"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiKeyAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "SolicitudVisita"
                ],
                "summary": "Actualizar una solicitud de visita por ID",
                "operationId": "4d1632e7699285f40c8183e1382bbc24",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la solicitud de visita",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SolicitudVisita"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Solicitud de visita actualizada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SolicitudVisita"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Solicitud de visita no encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Solicitud de visita no encontrada"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Errores de validación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "errors": {
                                            "type": "object",
                                            "additionalProperties": {
                                                "type": "string"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error al actualizar solicitud de visita",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Error al actualizar solicitud de visita: mensaje del error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiKeyAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "SolicitudVisita"
                ],
                "summary": "Eliminar una solicitud de visita por ID",
                "operationId": "e4042d1388d39e7f957beafa43921186",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la solicitud de visita",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Solicitud de visita eliminada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Solicitud de visita eliminada correctamente"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Solicitud de visita no encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Solicitud de visita no encontrada"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error al eliminar solicitud de visita",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Error al eliminar solicitud de visita: mensaje del error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiKeyAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "SolicitudVisita": {
                "title": "Solicitud de Visita",
                "description": "Modelo de Solicitud de Visita",
                "required": [
                    "persona_id",
                    "propiedad_id",
                    "fecha_visita"
                ],
                "properties": {
                    "persona_id": {
                        "type": "integer",
                        "example": 14
                    },
                    "propiedad_id": {
                        "type": "integer",
                        "example": 10
                    },
                    "fecha_visita": {
                        "type": "string",
                        "format": "date",
                        "example": "2024-09-12"
                    },
                    "comentarios": {
                        "type": "string",
                        "example": "Comentario sobre la visita"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "apiKeyAuth": {
                "type": "apiKey",
                "name": "X-RapidAPI-Key",
                "in": "header"
            }
        }
    },
...
    "security": [
        {
            "apiKeyAuth": {
                "type": "apiKey",
                "in": "header",
                "name": "X-RapidAPI-Key"
            }
        }
    ]
}